refactor(test): merge arcade multi-ops tests into single parameterized file
